+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 
-## 1.0.1 - 2025-11-14
+## 1.1.0 - 2026-07-01
+
+### Added
+
+- HSplitContainer offsets are now saved and fills entire area.
+
+### Changed
+
+- Updated with support for the latest Retrograde Image specification.
+
+## 1.0.1 - 2025-12-14
 
 ### Changed